Refactor MCV merging and Histogram tests outside CStatisticsTest
