+++ server/CMakeLists.txt
@@ -17,6 +17,8 @@ if (COMPILER_HAS_WNO_LONG_LONG)
     set_source_files_properties(src/unittest/dxl/statistics/CStatisticsTest.cpp
                                 src/unittest/dxl/statistics/CBucketTest.cpp
                                 src/unittest/dxl/statistics/CPointTest.cpp
+                                src/unittest/dxl/statistics/CHistogramTest.cpp
+                                src/unittest/dxl/statistics/CMCVTest.cpp
                                 src/unittest/dxl/statistics/CJoinCardinalityTest.cpp
                                 src/unittest/gpopt/base/CConstraintTest.cpp
                                 src/unittest/gpopt/metadata/CPartConstraintTest.cpp
@@ -325,6 +327,8 @@ add_orca_test(CParseHandlerTest)
 add_orca_test(CStatisticsTest)
 add_orca_test(CPointTest)
 add_orca_test(CBucketTest)
+add_orca_test(CHistogramTest)
+add_orca_test(CMCVTest)
 add_orca_test(CJoinCardinalityTest)
 add_orca_test(CTranslatorDXLToExprTest)
 add_orca_test(CTranslatorExprToDXLTest)
