+++ server/src/unittest/dxl/statistics/CBucketTest.cpp
@@ -1,12 +1,12 @@
 //---------------------------------------------------------------------------
 //	Greenplum Database
-//	Copyright (C) 2011 EMC Corp.
+//	Copyright (C) 2018 Pivotal Inc.
 //
 //	@filename:
 //		CBucketTest.cpp
 //
 //	@doc:
-//		Tests for CPoint
+//		Testing operations on histogram buckets
 //---------------------------------------------------------------------------
 
 #ifndef __STDC_CONSTANT_MACROS
