+++ server/src/unittest/dxl/statistics/CStatisticsTest.cpp
@@ -50,8 +50,6 @@ const CHAR *
 szInputDXLFileName = "../data/dxl/statistics/Basic-Statistics-Input.xml";
 const CHAR *
 szOutputDXLFileName = "../data/dxl/statistics/Basic-Statistics-Output.xml";
-const CHAR *
-szMCVSortExpectedFileName = "../data/dxl/statistics/MCV-Sort-Output.xml";
 
 const CHAR *
 szQuerySelect = "../data/dxl/statistics/SelectQuery.xml";
@@ -94,22 +92,16 @@ CStatisticsTest::EresUnittest()
 	// tests that use shared optimization context
 	CUnittest rgutSharedOptCtxt[] =
 		{
-		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_CHistogramInt4),
-		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_CHistogramBool),
 		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_CStatisticsBasic),
 		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_CStatisticsBasicsFromDXL),
 		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_CStatisticsBasicsFromDXLNumeric),
 		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_UnionAll),
 		// TODO,  Mar 18 2013 temporarily disabling the test
 		// GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_CStatisticsSelectDerivation),
-		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_Skew),
-		GPOS_UNITTEST_FUNC_ASSERT(CStatisticsTest::EresUnittest_CHistogramValid),
 		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_CStatisticsFilter),
 		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_CStatisticsFilterConj),
 		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_CStatisticsFilterDisj),
 		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_CStatisticsNestedPred),
-		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_SortInt4MCVs),
-		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_MergeHistMCV),
 		GPOS_UNITTEST_FUNC(CStatisticsTest::EresUnittest_CStatisticsAccumulateCard)
 		};
 
@@ -340,186 +332,7 @@ CStatisticsTest::EresUnittest_GbAggWithRepeatedGbCols()
 	return GPOS_FAILED;
 }
 
-//---------------------------------------------------------------------------
-//	@function:
-//		CStatisticsTest::EresUnittest_CHistogramInt4
-//
-//	@doc:
-//		Histogram of int4
-//
-//---------------------------------------------------------------------------
-GPOS_RESULT
-CStatisticsTest::EresUnittest_CHistogramInt4()
-{
-	// create memory pool
-	CAutoMemoryPool amp;
-	IMemoryPool *pmp = amp.Pmp();
-
-	// original histogram
-	CHistogram *phist = CCardinalityTestUtils::PhistExampleInt4(pmp);
-	CCardinalityTestUtils::PrintHist(pmp, "phist", phist);
-
-	// test edge case of PbucketGreaterThan
-	CPoint *ppoint0 = CTestUtils::PpointInt4(pmp, 9);
-	CHistogram *phist0 = phist->PhistFilter(pmp, CStatsPred::EstatscmptG, ppoint0);
-	CCardinalityTestUtils::PrintHist(pmp, "phist0", phist0);
-	GPOS_RTL_ASSERT(phist0->UlBuckets() == 9);
-
-	CPoint *ppoint1 = CTestUtils::PpointInt4(pmp, 35);
-	CHistogram *phist1 = phist->PhistFilter(pmp, CStatsPred::EstatscmptL, ppoint1);
-	CCardinalityTestUtils::PrintHist(pmp, "phist1", phist1);
-	GPOS_RTL_ASSERT(phist1->UlBuckets() == 4);
-
-	// edge case where point is equal to upper bound
-	CPoint *ppoint2 = CTestUtils::PpointInt4(pmp, 50);
-	CHistogram *phist2 = phist->PhistFilter(pmp, CStatsPred::EstatscmptL,ppoint2);
-	CCardinalityTestUtils::PrintHist(pmp, "phist2", phist2);
-	GPOS_RTL_ASSERT(phist2->UlBuckets() == 5);
-
-	// equality check
-	CPoint *ppoint3 = CTestUtils::PpointInt4(pmp, 100);
-	CHistogram *phist3 = phist->PhistFilter(pmp, CStatsPred::EstatscmptEq, ppoint3);
-	CCardinalityTestUtils::PrintHist(pmp, "phist3", phist3);
-	GPOS_RTL_ASSERT(phist3->UlBuckets() == 1);
-
-	// normalized output after filter
-	CPoint *ppoint4 = CTestUtils::PpointInt4(pmp, 100);
-	CDouble dScaleFactor(0.0);
-	CHistogram *phist4 = phist->PhistFilterNormalized(pmp, CStatsPred::EstatscmptEq, ppoint4, &dScaleFactor);
-	CCardinalityTestUtils::PrintHist(pmp, "phist4", phist4);
-	GPOS_RTL_ASSERT(phist4->FValid());
-
-	// lasj
-	CHistogram *phist5 = phist->PhistLASJ(pmp, CStatsPred::EstatscmptEq, phist2);
-	CCardinalityTestUtils::PrintHist(pmp, "phist5", phist5);
-	GPOS_RTL_ASSERT(phist5->UlBuckets() == 5);
-
-	// inequality check
-	CHistogram *phist6 = phist->PhistFilter(pmp, CStatsPred::EstatscmptNEq, ppoint2);
-	CCardinalityTestUtils::PrintHist(pmp, "phist6", phist6);
-	GPOS_RTL_ASSERT(phist6->UlBuckets() == 10);
-
-	// histogram with null fraction and remaining tuples
-	CHistogram *phist7 = PhistExampleInt4Remain(pmp);
-	CCardinalityTestUtils::PrintHist(pmp, "phist7", phist7);
-	CPoint *ppoint5 = CTestUtils::PpointInt4(pmp, 20);
-
-	// equality check, hitting remaining tuples
-	CHistogram *phist8 = phist7->PhistFilter(pmp, CStatsPred::EstatscmptEq, ppoint3);
-	GPOS_RTL_ASSERT(fabs((phist8->DFrequency() - 0.2).DVal()) < CStatistics::DEpsilon);
-	GPOS_RTL_ASSERT(fabs((phist8->DDistinct() - 1.0).DVal()) < CStatistics::DEpsilon);
-
-	// greater than, hitting remaining tuples
-	CHistogram *phist9 = phist7->PhistFilter(pmp, CStatsPred::EstatscmptG, ppoint1);
-	CCardinalityTestUtils::PrintHist(pmp, "phist9", phist9);
-	GPOS_RTL_ASSERT(fabs((phist9->DFrequency() - 0.26).DVal()) < CStatistics::DEpsilon);
-	GPOS_RTL_ASSERT(fabs((phist9->DDistinct() - 1.8).DVal()) < CStatistics::DEpsilon);
-
-	// equality join, hitting remaining tuples
-	CHistogram *phist10 = phist7->PhistJoin(pmp, CStatsPred::EstatscmptEq, phist7);
-	GPOS_RTL_ASSERT(phist10->UlBuckets() == 5);
-	GPOS_RTL_ASSERT(fabs((phist10->DDistinctRemain() - 2.0).DVal()) < CStatistics::DEpsilon);
-	GPOS_RTL_ASSERT(fabs((phist10->DFreqRemain() - 0.08).DVal()) < CStatistics::DEpsilon);
-
-	// clean up
-	ppoint0->Release();
-	ppoint1->Release();
-	ppoint2->Release();
-	ppoint3->Release();
-	ppoint4->Release();
-	ppoint5->Release();
-	GPOS_DELETE(phist);
-	GPOS_DELETE(phist0);
-	GPOS_DELETE(phist1);
-	GPOS_DELETE(phist2);
-	GPOS_DELETE(phist3);
-	GPOS_DELETE(phist4);
-	GPOS_DELETE(phist5);
-	GPOS_DELETE(phist6);
-	GPOS_DELETE(phist7);
-	GPOS_DELETE(phist8);
-	GPOS_DELETE(phist9);
-	GPOS_DELETE(phist10);
-
-	return GPOS_OK;
-}
-
-//---------------------------------------------------------------------------
-//	@function:
-//		CStatisticsTest::EresUnittest_CHistogramBool
-//
-//	@doc:
-//		Histogram on bool
-//
-//---------------------------------------------------------------------------
-GPOS_RESULT
-CStatisticsTest::EresUnittest_CHistogramBool()
-{
-	// create memory pool
-	CAutoMemoryPool amp;
-	IMemoryPool *pmp = amp.Pmp();
-
-	// generate histogram of the form [false, false), [true,true)
-	DrgPbucket *pdrgppbucket = GPOS_NEW(pmp) DrgPbucket(pmp);
-	CBucket *pbucketFalse = CCardinalityTestUtils::PbucketSingletonBoolVal(pmp, false, 0.1);
-	CBucket *pbucketTrue = CCardinalityTestUtils::PbucketSingletonBoolVal(pmp, false, 0.9);
-	pdrgppbucket->Append(pbucketFalse);
-	pdrgppbucket->Append(pbucketTrue);
-	CHistogram *phist =  GPOS_NEW(pmp) CHistogram(pdrgppbucket);
-
-	// equality check
-	CPoint *ppoint1 = CTestUtils::PpointBool(pmp, false);
-	CDouble dScaleFactor(0.0);
-	CHistogram *phist1 = phist->PhistFilterNormalized(pmp, CStatsPred::EstatscmptEq, ppoint1, &dScaleFactor);
-	CCardinalityTestUtils::PrintHist(pmp, "phist1", phist1);
-	GPOS_RTL_ASSERT(phist1->UlBuckets() == 1);
-
-	// clean up
-	ppoint1->Release();
-	GPOS_DELETE(phist);
-	GPOS_DELETE(phist1);
-
-	return GPOS_OK;
-}
-
-//---------------------------------------------------------------------------
-//	@function:
-//		CStatisticsTest::PhistExampleInt4Remain
-//
-//	@doc:
-//		Generates example int histogram having tuples not covered by buckets,
-//		including null fraction and nDistinctRemain
-//
-//---------------------------------------------------------------------------
-CHistogram*
-CStatisticsTest::PhistExampleInt4Remain
-	(
-	IMemoryPool *pmp
-	)
-{
-	// generate histogram of the form [0, 0], [10, 10], [20, 20] ...
-	DrgPbucket *pdrgppbucket = GPOS_NEW(pmp) DrgPbucket(pmp);
-	for (ULONG ulIdx = 0; ulIdx < 5; ulIdx++)
-	{
-		INT iLower = INT(ulIdx * 10);
-		INT iUpper = iLower;
-		CDouble dFrequency(0.1);
-		CDouble dDistinct(1.0);
-		CBucket *pbucket = CCardinalityTestUtils::PbucketIntegerClosedLowerBound(pmp, iLower, iUpper, dFrequency, dDistinct);
-		pdrgppbucket->Append(pbucket);
-	}
-
-	return GPOS_NEW(pmp) CHistogram(pdrgppbucket, true, 0.1 /*dNullFreq*/, 2.0 /*dDistinctRemain*/, 0.4 /*dFreqRemain*/);
-}
-
-//---------------------------------------------------------------------------
-//	@function:
-//		CStatisticsTest::PhistExampleInt4Dim
-//
-//	@doc:
-//		Generates example int histogram corresponding to dimension table
-//
-//---------------------------------------------------------------------------
+// generates example int histogram corresponding to dimension table
 CHistogram*
 CStatisticsTest::PhistExampleInt4Dim
 	(
@@ -2012,42 +1825,6 @@ CStatisticsTest::Pdrgpstatspred2
 
 //---------------------------------------------------------------------------
 //	@function:
-//		CStatisticsTest::EresUnittest_CHistogramValid
-//
-//	@doc:
-//		Check for well-formed histogram. Expected to hit an assert.
-//
-//---------------------------------------------------------------------------
-GPOS_RESULT
-CStatisticsTest::EresUnittest_CHistogramValid()
-{
-	// create memory pool
-	CAutoMemoryPool amp;
-	IMemoryPool *pmp = amp.Pmp();
-
-	DrgPbucket *pdrgppbucket = GPOS_NEW(pmp) DrgPbucket(pmp);
-
-	// generate histogram of the form [0, 10), [9, 20)
-	CBucket *pbucket1 = CCardinalityTestUtils::PbucketIntegerClosedLowerBound(pmp, 0, 10, 0.1, 2.0);
-	pdrgppbucket->Append(pbucket1);
-	CBucket *pbucket2 = CCardinalityTestUtils::PbucketIntegerClosedLowerBound(pmp, 9, 20, 0.1, 2.0);
-	pdrgppbucket->Append(pbucket2);
-
-	// original histogram
-	CHistogram *phist =  GPOS_NEW(pmp) CHistogram(pdrgppbucket);
-
-	// create an auto object
-	CAutoP<CHistogram> ahist;
-	ahist = phist;
-
-	GPOS_RTL_ASSERT(phist->FValid() && "Histogram must be well formed");
-
-	return GPOS_FAILED;
-}
-
-
-//---------------------------------------------------------------------------
-//	@function:
 //		CStatisticsTest::StatsFilterInt4
 //
 //	@doc:
@@ -2208,300 +1985,6 @@ CStatisticsTest::EresUnittest_CStatisticsSelectDerivation()
 			szPlanSelect,
 			true // ignore mismatch in output dxl due to column id differences
 			);
-}
-
-
-//---------------------------------------------------------------------------
-//	@function:
-//		CStatisticsTest::EresUnittest_Skew
-//
-//	@doc:
-//		Basis skew test
-//
-//---------------------------------------------------------------------------
-GPOS_RESULT
-CStatisticsTest::EresUnittest_Skew()
-{
-	// create memory pool
-	CAutoMemoryPool amp;
-	IMemoryPool *pmp = amp.Pmp();
-
-	CBucket *pbucket1 = CCardinalityTestUtils::PbucketIntegerClosedLowerBound(pmp, 1, 100, CDouble(0.6), CDouble(100.0));
-	CBucket *pbucket2 = CCardinalityTestUtils::PbucketIntegerClosedLowerBound(pmp, 101, 200, CDouble(0.2), CDouble(100.0));
-	CBucket *pbucket3 = CCardinalityTestUtils::PbucketIntegerClosedLowerBound(pmp, 201, 300, CDouble(0.2), CDouble(100.0));
-	CBucket *pbucket4 = CCardinalityTestUtils::PbucketIntegerClosedLowerBound(pmp, 301, 400, CDouble(0.2), CDouble(100.0));
-	CBucket *pbucket5 = CCardinalityTestUtils::PbucketIntegerClosedLowerBound(pmp, 401, 500, CDouble(0.2), CDouble(100.0));
-	CBucket *pbucket6 = CCardinalityTestUtils::PbucketIntegerClosedLowerBound(pmp, 501, 600, CDouble(0.2), CDouble(100.0));
-	CBucket *pbucket7 = CCardinalityTestUtils::PbucketIntegerClosedLowerBound(pmp, 601, 700, CDouble(0.2), CDouble(100.0));
-	CBucket *pbucket8 = CCardinalityTestUtils::PbucketIntegerClosedLowerBound(pmp, 701, 800, CDouble(0.2), CDouble(100.0));
-
-	DrgPbucket *pdrgppbucket1 = GPOS_NEW(pmp) DrgPbucket(pmp);
-	pdrgppbucket1->Append(pbucket1);
-	pdrgppbucket1->Append(pbucket2);
-	pdrgppbucket1->Append(pbucket3);
-	CHistogram *phist1 =  GPOS_NEW(pmp) CHistogram(pdrgppbucket1);
-
-	DrgPbucket *pdrgppbucket2 = GPOS_NEW(pmp) DrgPbucket(pmp);
-	pdrgppbucket2->Append(pbucket4);
-	pdrgppbucket2->Append(pbucket5);
-	pdrgppbucket2->Append(pbucket6);
-	pdrgppbucket2->Append(pbucket7);
-	pdrgppbucket2->Append(pbucket8);
-	CHistogram *phist2 =  GPOS_NEW(pmp) CHistogram(pdrgppbucket2);
-	GPOS_ASSERT(phist1->DSkew() > phist2->DSkew());
-
-	{
-		CAutoTrace at(pmp);
-		phist1->OsPrint(at.Os());
-		phist2->OsPrint(at.Os());
-	}
-
-	GPOS_DELETE(phist1);
-	GPOS_DELETE(phist2);
-
-	return GPOS_OK;
-}
-
-//---------------------------------------------------------------------------
-//	@function:
-//		CStatisticsTest::EresUnittest_SortInt4MCVs
-//
-//	@doc:
-//		Test sorting of Int4 MCVs and their associated frequencies
-//
-//---------------------------------------------------------------------------
-GPOS_RESULT
-CStatisticsTest::EresUnittest_SortInt4MCVs()
-{
-	// create memory pool
-	CAutoMemoryPool amp;
-	IMemoryPool *pmp = amp.Pmp();
-
-	CMDAccessor *pmda = COptCtxt::PoctxtFromTLS()->Pmda();
-	CMDIdGPDB *pmdid = GPOS_NEW(pmp) CMDIdGPDB(CMDIdGPDB::m_mdidInt4);
-	const IMDType *pmdtype = pmda->Pmdtype(pmdid);
-
-	// create three integer MCVs
-	CPoint *ppoint1 = CTestUtils::PpointInt4(pmp, 5);
-	CPoint *ppoint2 = CTestUtils::PpointInt4(pmp, 1);
-	CPoint *ppoint3 = CTestUtils::PpointInt4(pmp, 10);
-	DrgPdatum *pdrgpdatumMCV = GPOS_NEW(pmp) DrgPdatum(pmp);
-	IDatum *pdatum1 = ppoint1->Pdatum();
-	IDatum *pdatum2 = ppoint2->Pdatum();
-	IDatum *pdatum3 = ppoint3->Pdatum();
-	pdatum1->AddRef();
-	pdatum2->AddRef();
-	pdatum3->AddRef();
-	pdrgpdatumMCV->Append(pdatum1);
-	pdrgpdatumMCV->Append(pdatum2);
-	pdrgpdatumMCV->Append(pdatum3);
-
-	// create corresponding frequencies
-	DrgPdouble *pdrgpdFreq = GPOS_NEW(pmp) DrgPdouble(pmp);
-	// in GPDB, MCVs are stored in descending frequencies
-	pdrgpdFreq->Append(GPOS_NEW(pmp) CDouble(0.4));
-	pdrgpdFreq->Append(GPOS_NEW(pmp) CDouble(0.2));
-	pdrgpdFreq->Append(GPOS_NEW(pmp) CDouble(0.1));
-
-	// exercise MCV sorting function
-	CHistogram *phistMCV = CStatisticsUtils::PhistTransformMCV
-								(
-								pmp,
-								pmdtype,
-								pdrgpdatumMCV,
-								pdrgpdFreq,
-								pdrgpdatumMCV->UlLength()
-								);
-
-	// create hash map from colid -> histogram
-	HMUlHist *phmulhist = GPOS_NEW(pmp) HMUlHist(pmp);
-
-	// generate int histogram for column 1
-	phmulhist->FInsert(GPOS_NEW(pmp) ULONG(1), phistMCV);
-
-	// column width for int4
-	HMUlDouble *phmuldoubleWidth = GPOS_NEW(pmp) HMUlDouble(pmp);
-	phmuldoubleWidth->FInsert(GPOS_NEW(pmp) ULONG(1), GPOS_NEW(pmp) CDouble(4.0));
-
-	CStatistics *pstats = GPOS_NEW(pmp) CStatistics
-									(
-									pmp,
-									phmulhist,
-									phmuldoubleWidth,
-									1000.0 /* dRows */,
-									false /* fEmpty */
-									);
-
-	// put stats object in an array in order to serialize
-	DrgPstats *pdrgpstats = GPOS_NEW(pmp) DrgPstats(pmp);
-	pdrgpstats->Append(pstats);
-
-	// serialize stats object
-	CWStringDynamic *pstrOutput = CDXLUtils::PstrSerializeStatistics(pmp, pmda, pdrgpstats, true, true);
-	GPOS_TRACE(pstrOutput->Wsz());
-
-	// get expected output
-	CWStringDynamic str(pmp);
-	COstreamString oss(&str);
-	CHAR *szDXLExpected = CDXLUtils::SzRead(pmp, szMCVSortExpectedFileName);
-	CWStringDynamic dstrExpected(pmp);
-	dstrExpected.AppendFormat(GPOS_WSZ_LIT("%s"), szDXLExpected);
-
-	GPOS_RESULT eres = CTestUtils::EresCompare
-								(
-								oss,
-								pstrOutput,
-								&dstrExpected,
-								false // mismatch will not be ignored
-								);
-
-	// cleanup
-	GPOS_DELETE(pstrOutput);
-	GPOS_DELETE_ARRAY(szDXLExpected);
-	pdrgpdatumMCV->Release();
-	pdrgpdFreq->Release();
-	pdrgpstats->Release();
-	ppoint1->Release();
-	ppoint2->Release();
-	ppoint3->Release();
-	pmdid->Release();
-
-	return eres;
-}
-
-//---------------------------------------------------------------------------
-//	@function:
-//		CStatisticsTest::EresUnittest_MergeHistMCV
-//
-//	@doc:
-//		Test merging MCVs and histogram
-//
-//---------------------------------------------------------------------------
-GPOS_RESULT
-CStatisticsTest::EresUnittest_MergeHistMCV()
-{
-	SMergeTestElem rgMergeTestElem[] =
-	{
-		{
-		"../data/dxl/statistics/Merge-Input-MCV-Int.xml",
-		"../data/dxl/statistics/Merge-Input-Histogram-Int.xml",
-		"../data/dxl/statistics/Merge-Output-Int.xml"
-		},
-
-		{
-		"../data/dxl/statistics/Merge-Input-MCV-Numeric.xml",
-		"../data/dxl/statistics/Merge-Input-Histogram-Numeric.xml",
-		"../data/dxl/statistics/Merge-Output-Numeric.xml"
-		},
-
-		{
-		"../data/dxl/statistics/Merge-Input-MCV-BPChar.xml",
-		"../data/dxl/statistics/Merge-Input-Histogram-BPChar.xml",
-		"../data/dxl/statistics/Merge-Output-BPChar.xml"
-		}
-	};
-
-	// create memory pool
-	CAutoMemoryPool amp;
-	IMemoryPool *pmp = amp.Pmp();
-
-	ULONG ulLen = GPOS_ARRAY_SIZE(rgMergeTestElem);
-	for (ULONG ul = 0; ul < ulLen; ul++)
-	{
-		// read input MCVs DXL file
-		CHAR *szDXLInputMCV = CDXLUtils::SzRead(pmp, rgMergeTestElem[ul].szInputMCVFile);
-		// read input histogram DXL file
-		CHAR *szDXLInputHist = CDXLUtils::SzRead(pmp, rgMergeTestElem[ul].szInputHistFile);
-
-		GPOS_CHECK_ABORT;
-
-		CMDAccessor *pmda = COptCtxt::PoctxtFromTLS()->Pmda();
-
-		// parse the stats objects
-		DrgPdxlstatsderrel *pdrgpdxlstatsderrelMCV = CDXLUtils::PdrgpdxlstatsderrelParseDXL(pmp, szDXLInputMCV, NULL);
-		DrgPdxlstatsderrel *pdrgpdxlstatsderrelHist = CDXLUtils::PdrgpdxlstatsderrelParseDXL(pmp, szDXLInputHist, NULL);
-
-		GPOS_CHECK_ABORT;
-
-		CDXLStatsDerivedRelation *pdxlstatsderrelMCV = (*pdrgpdxlstatsderrelMCV)[0];
-		const DrgPdxlstatsdercol *pdrgpdxlstatsdercolMCV = pdxlstatsderrelMCV->Pdrgpdxlstatsdercol();
-		CDXLStatsDerivedColumn *pdxlstatsdercolMCV = (*pdrgpdxlstatsdercolMCV)[0];
-		DrgPbucket *pdrgppbucketMCV = CDXLUtils::Pdrgpbucket(pmp, pmda, pdxlstatsdercolMCV);
-		CHistogram *phistMCV =  GPOS_NEW(pmp) CHistogram(pdrgppbucketMCV);
-
-		CDXLStatsDerivedRelation *pdxlstatsderrelHist = (*pdrgpdxlstatsderrelHist)[0];
-		const DrgPdxlstatsdercol *pdrgpdxlstatsdercolHist = pdxlstatsderrelHist->Pdrgpdxlstatsdercol();
-		CDXLStatsDerivedColumn *pdxlstatsdercolHist = (*pdrgpdxlstatsdercolHist)[0];
-		DrgPbucket *pdrgppbucketHist = CDXLUtils::Pdrgpbucket(pmp, pmda, pdxlstatsdercolHist);
-		CHistogram *phistHist =  GPOS_NEW(pmp) CHistogram(pdrgppbucketHist);
-
-		GPOS_CHECK_ABORT;
-
-		// exercise the merge
-		CHistogram *phistMerged = CStatisticsUtils::PhistMergeMcvHist(pmp, phistMCV, phistHist);
-
-		// create hash map from colid -> histogram
-		HMUlHist *phmulhist = GPOS_NEW(pmp) HMUlHist(pmp);
-
-		// generate int histogram for column 1
-		ULONG ulColId = pdxlstatsdercolMCV->UlColId();
-		phmulhist->FInsert(GPOS_NEW(pmp) ULONG(ulColId), phistMerged);
-
-		// column width for int4
-		HMUlDouble *phmuldoubleWidth = GPOS_NEW(pmp) HMUlDouble(pmp);
-		CDouble dWidth = pdxlstatsdercolMCV->DWidth();
-		phmuldoubleWidth->FInsert(GPOS_NEW(pmp) ULONG(ulColId), GPOS_NEW(pmp) CDouble(dWidth));
-
-		CStatistics *pstats = GPOS_NEW(pmp) CStatistics
-										(
-										pmp,
-										phmulhist,
-										phmuldoubleWidth,
-										pdxlstatsderrelMCV->DRows(),
-										pdxlstatsderrelMCV->FEmpty()
-										);
-
-		// put stats object in an array in order to serialize
-		DrgPstats *pdrgpstats = GPOS_NEW(pmp) DrgPstats(pmp);
-		pdrgpstats->Append(pstats);
-
-		// serialize stats object
-		CWStringDynamic *pstrOutput = CDXLUtils::PstrSerializeStatistics(pmp, pmda, pdrgpstats, true, true);
-		GPOS_TRACE(pstrOutput->Wsz());
-
-		// get expected output
-		CWStringDynamic str(pmp);
-		COstreamString oss(&str);
-		CHAR *szDXLExpected = CDXLUtils::SzRead(pmp, rgMergeTestElem[ul].szMergedFile);
-		CWStringDynamic dstrExpected(pmp);
-		dstrExpected.AppendFormat(GPOS_WSZ_LIT("%s"), szDXLExpected);
-
-		GPOS_RESULT eres = CTestUtils::EresCompare
-									(
-									oss,
-									pstrOutput,
-									&dstrExpected,
-									false // mismatch will not be ignored
-									);
-
-		// cleanup
-		GPOS_DELETE_ARRAY(szDXLInputMCV);
-		GPOS_DELETE_ARRAY(szDXLInputHist);
-		GPOS_DELETE_ARRAY(szDXLExpected);
-		GPOS_DELETE(pstrOutput);
-		pdrgpdxlstatsderrelMCV->Release();
-		pdrgpdxlstatsderrelHist->Release();
-		GPOS_DELETE(phistMCV);
-		GPOS_DELETE(phistHist);
-		pdrgpstats->Release();
-
-		if (GPOS_OK != eres)
-		{
-			return eres;
-		}
-	}
-
-	return GPOS_OK;
 }
 
 //---------------------------------------------------------------------------
